Conflicts:
	target/m2e-wtp/web-resources/META-INF/maven/AD/Store/pom.properties


	y cambios de long el numtarjeta, formato de editar y crear cliente en
las vistas y cambios en los import, schema y base de datos

+++ src/ad/store/dao/UserDao.java
@@ -8,7 +8,7 @@ import ad.store.entity.Cliente;
 public interface UserDao extends GenericDao<Cliente>{
 
 	public Cliente altaCliente(String nombreUsuario, String direccionEnvio, String password, String nombre,
-			String apellidos, String email, Date fechaNa, String banco, int numTarjeta, String titular,
+			String apellidos, String email, Date fechaNa, String banco, long numTarjeta, String titular,
 			int codigoS, String direccionFa);
 	
 	public Cliente obtenerCliente(long idCliente);
