+++ src/ad/store/service/UserServiceImpl.java
@@ -35,7 +35,7 @@ public class UserServiceImpl implements UserService {
 	
 	@Override
 	public Cliente altaCliente(String nombreUsuario, String direccionEnvio, String password, String nombre,
-			String apellidos, String email, Date fechaNa, String banco, int numTarjeta, String titular,
+			String apellidos, String email, Date fechaNa, String banco, long numTarjeta, String titular,
 			int codigoS, String direccionFa) {
 		Cliente cliente = new Cliente();
 		cliente.setNombreUsuario(nombreUsuario);
